mandatory done and normed bonus 5/9

+++ libft/ft_memchr.c
@@ -12,24 +12,24 @@
 
 #include <string.h>
 
-void *ft_memchr(const void *s, int c, size_t n)
-    {
-        long unsigned int i;
-        int j;
-        const char *s1;
+void	*ft_memchr(const void *s, int c, size_t n)
+{
+	size_t			i;
+	int				j;
+	const char		*s1;
 
-        i = 0;
-        j = 0;
-        s1 = s;
-        while(i < (n - 1))
-            {
-                if(s1[j] == c)
-                    return((char *)s + i);
-                else
-                {
-                    i++;
-                    j++;
-                }
-            }
-        return(0);
-    }
+	i = 0;
+	j = 0;
+	s1 = s;
+	while (i < (n - 1))
+	{
+		if (s1[j] == c)
+			return ((char *)s + i);
+		else
+		{
+			i++;
+			j++;
+		}
+	}
+	return (0);
+}
